Update machine_frame.dm (#32293)

+++ code/game/machinery/machine_frame.dm
@@ -249,6 +249,8 @@
 		return TRUE
 	I.play_tool_sound(src)
 	var/obj/machinery/new_machine = new circuit.build_path(loc)
+	if(QDELETED(new_machine))
+		return TRUE
 	new_machine.on_construction()
 	for(var/obj/O in new_machine.component_parts)
 		qdel(O)
